+++ src/App.css
@@ -1,6 +1,5 @@
 .App {
   text-align: center;
-  height: 100%;
 }
 
 .App-logo {
