+++ src/components/Login/Login.scss
@@ -1,0 +1,99 @@
+.label{
+  font-family: Nunito Sans;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 14px;
+  line-height: 19px;
+  color: #000000;
+  mix-blend-mode: normal;
+  opacity: 0.9;
+}
+
+.header{
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 24px;
+  line-height: 33px;
+  text-align: center;
+  /* identical to box height */
+  color: #FFFFFF;
+}
+
+.required:after {
+  content:"*";
+  color: #FF3E06;
+}
+
+.body{
+  background: linear-gradient(180deg, #0F2D58 0%, #2C5797 55.91%);
+}
+
+.form-input{
+  background: #F6F6F6;
+  border-radius: 5px;
+  height: 25px;
+  width: 405px;
+}
+
+.fieldset{
+  background: #FFFFFF;
+  box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.147044);
+  border-radius: 8px;
+  width: 540px;
+  left: 450px;
+  top: 217px;
+  align-self: center;
+
+}
+
+.forgotpwd{
+  font-family: Nunito Sans;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 14px;
+  line-height: 19px;
+  /* identical to box height */
+
+  color: #FF3C00;
+
+  mix-blend-mode: normal;
+  opacity: 0.9;
+}
+
+
+.btnLogin
+{
+  border-color:#FF3C00;
+  border-radius : 5px;
+  background-color : #FFFFFF;
+  color: #FF3C00;
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 20px;
+    height: 48px;
+}
+
+.btnSignUp
+{
+  background-color : #FF3C00;
+  color:#FFFFFF;
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 20px;
+    height: 48px;
+}
+
+.member{
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 14px;
+  line-height: 19px;
+
+  color: #000000;
+}
