+++ src/index.css
@@ -3,7 +3,3 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
-
-html, body, #app, #app>div {
-  height: 100%
-}
